Add an “All” tab to releases list
Using it myself, I find that the most natural starting point for a
curious person is to start at whatever the page starts at and scroll
down. This gives a default that appeals to the new, interested person
while also giving the option to view only free or pro updates.