Add Dashify/Pro version to all release titles
This allows the title to make sense when standing alone. For example,
“Dashify 1.9.0: Improved search” has more meaning than just “Improved
search”, especially in places with little context like the footer recent
releases or Google search results.

This also makes all the release titles more consistent, as some already
had the plugin name and version when there wasn’t anything notable to include
in the title.