Add URL query param for Releases tab selection
This will allow anyone to link to a specific set of releases, Free or
Pro. This is useful for example for linking from the “View version x
details” link from the plugins list in the WordPress admin. For the Pro,
we can filter to just the Pro ones, to create a Pro changelog.